+++ mockbird-server/src/main/java/com/mockbird/entity/Project.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import javax.validation.constraints.NotBlank;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,6 +15,7 @@ public class Project {
     @TableId(type = IdType.AUTO)
     private Long id;
 
+    @NotBlank(message = "项目名称不能为空")
     private String name;
 
     private String description;